requirements.txt
fastmcp==0.2.0
python-dotenv
pydantic
markdownify
httpx
readabilipy
uvicorn
Pillow
streamlit
